首页蓝牙数据解析

+++ bluebao/Modules/工具箱/MyTool.m
@@ -159,7 +159,7 @@
 
 +(NSString *) getStringFormstr:(NSString *)string withNumber:(NSInteger)number{
     
-    NSString * str = [NSString stringWithFormat:@"%@%ld",string,number];
+    NSString * str = [NSString stringWithFormat:@"%@%ld",string,(long)number];
     return str;
 }
